Renamed login components folder to auth
