+++ frontend/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "@/components/login/login-form"
+import LoginForm from "@/components/auth/login-form"
 
 export default function LoginPage() {
   return (
