+++ frontend/src/app/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignupForm from "@/components/login/signup-form"
+import SignupForm from "@/components/auth/signup-form"
 
 
 export default function SignupPage () {
